+++ test/selic.spec.ts
@@ -1,6 +1,6 @@
 import * as nock from 'nock'
 import { BCB_API, BCB_SELIC_PATH, CDI_SCORE, POUPANCA_PERCENT } from '../src/constants'
-import Selic from '../src/selic'
+import { Selic } from '../src/selic'
 
 describe('Selic', () => {
   describe('.scrapRates', () => {
@@ -63,7 +63,7 @@ describe('Selic', () => {
     });
   });
 
-  describe('.getSelic', () => {
+  describe('.getSelicRate', () => {
     test('returns the selic rates when bcb scraps succeed', async () => {
       const fakeSelic = 10.1111;
       const fakeData = JSON.stringify([{ valor: fakeSelic }]);
@@ -72,7 +72,7 @@ describe('Selic', () => {
       
       const selic = new Selic();
 
-      const value = await selic.getSelic();
+      const value = await selic.getSelicRate();
       const expected = Number(Number(fakeSelic).toFixed(2));
       expect(value).toBe(expected);
     });
@@ -84,7 +84,7 @@ describe('Selic', () => {
 
       const selic = new Selic();
       
-      await expect(selic.getSelic()).rejects.toThrow('Parse error');
+      await expect(selic.getSelicRate()).rejects.toThrow('Parse error');
     });
     
     test('raises error when bcb scraps fail', async () => {
@@ -92,7 +92,73 @@ describe('Selic', () => {
 
       const selic = new Selic();
       
-      await expect(selic.getSelic()).rejects.toThrow('Request error');
+      await expect(selic.getSelicRate()).rejects.toThrow('Request error');
+    });
+  });
+
+  describe('.getCdiRate', () => {
+    test('returns the selic rates when bcb scraps succeed', async () => {
+      const fakeSelic = 10.1111;
+      const fakeData = JSON.stringify([{ valor: fakeSelic }]);
+
+      nock(BCB_API).get(BCB_SELIC_PATH).reply(200, fakeData);
+      
+      const selic = new Selic();
+
+      const value = await selic.getCdiRate();
+      const expected = Number(Number(fakeSelic - CDI_SCORE).toFixed(2));
+      expect(value).toBe(expected);
+    });
+
+    test('raises error when bcb scraps parse fail', async () => {
+      const fakeData = '<>INVALID</>';
+
+      nock(BCB_API).get(BCB_SELIC_PATH).reply(200, fakeData);
+
+      const selic = new Selic();
+      
+      await expect(selic.getCdiRate()).rejects.toThrow('Parse error');
+    });
+    
+    test('raises error when bcb scraps fail', async () => {
+      nock(BCB_API).get(BCB_SELIC_PATH).replyWithError('Api fails');
+
+      const selic = new Selic();
+      
+      await expect(selic.getCdiRate()).rejects.toThrow('Request error');
+    });
+  });
+
+  describe('.getPoupancaRate', () => {
+    test('returns the selic rates when bcb scraps succeed', async () => {
+      const fakeSelic = 10.1111;
+      const fakeData = JSON.stringify([{ valor: fakeSelic }]);
+
+      nock(BCB_API).get(BCB_SELIC_PATH).reply(200, fakeData);
+      
+      const selic = new Selic();
+
+      const value = await selic.getPoupancaRate();
+      const expected = Number(Number((fakeSelic / 100) * POUPANCA_PERCENT).toFixed(2));
+      expect(value).toBe(expected);
+    });
+
+    test('raises error when bcb scraps parse fail', async () => {
+      const fakeData = '<>INVALID</>';
+
+      nock(BCB_API).get(BCB_SELIC_PATH).reply(200, fakeData);
+
+      const selic = new Selic();
+      
+      await expect(selic.getPoupancaRate()).rejects.toThrow('Parse error');
+    });
+    
+    test('raises error when bcb scraps fail', async () => {
+      nock(BCB_API).get(BCB_SELIC_PATH).replyWithError('Api fails');
+
+      const selic = new Selic();
+      
+      await expect(selic.getPoupancaRate()).rejects.toThrow('Request error');
     });
   });